requirements.txt
meshtastic
pubsub
pyephem
requests
maidenhead
beautifulsoup4
dadjokes
geopy
schedule
wikipedia
googlesearch-python
google-auth
lxml
